Beginnings of discv5 wire protocol
